requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
python-dotenv
passlib[bcrypt]
PyJWT
requests
pydantic